requirements.txt
pytest==8.3.3
pytest-cov==5.0.0
pylint==3.3.1
black==24.10.0
fire==0.7.0
spacy==2.3.9
es_core_news_sm==2.3.1
#torch==1.13.0+cpu --index-url https://download.pytorch.org/whl/cpu

langchain-community
sentence-transformers
#langchain==0.3.7
#langchain-core==0.3.17
#langchain-text-splitters==0.3.2
pdfminer.six
#pypdf
streamlit==1.40.1
pi_heif
# unstructured
# unstructured[local-inference]
unstructured==0.16.5
unstructured-client==0.27.0
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
# watchdog
qdrant-client==1.12.1
qdrant-client[fastembed]==1.12.1
#watchdog==6.0.0
#langchain_qdrant
# json
# os
torch --pre --index-url https://download.pytorch.org/whl/nightly/cpu/torch_nightly.html
#langchain-huggingface>=0.2.2
langchain-huggingface
PyPDF2
ragas
#textblob
language_tool_python
